refactor

+++ src/main/java/com/codecool/models/User.java
@@ -9,23 +9,23 @@ import java.util.Set;
 public class User {
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
     private String name;
     private String surname;
     @ManyToMany(cascade = { CascadeType.ALL })
     private Set<Car> cars = new HashSet<>();
 
-    public User(int id, String name, String surname) {
+    public User(long id, String name, String surname) {
         this.id = id;
         this.name = name;
         this.surname = surname;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
